# contracts/openapi/yumpoo-v1.yaml
openapi: 3.0.3
info:
  title: Yumpoo API
  version: 1.0.0
  description: |
    Yumpoo 一期 HTTP API 的唯一静态契约源。

    所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共
    schema、header、parameter、response 和 golden example；业务 paths 在对应模块
    实现时按兼容新增原则加入。
servers:
  - url: /api/v1
    description: Yumpoo v1 business API
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: 查询调用人可见的 Project
      description: lifecycle 省略时返回 DRAFT 与 ACTIVE；分页和总数使用相同权限及筛选谓词。
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: workspaceId, in: query, deprecated: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: projectType, in: query, deprecated: true, schema: { $ref: '#/components/schemas/ProjectType' } }
        - { name: query, in: query, schema: { type: string, maxLength: 80 } }
        - name: projectTypes
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: '#/components/schemas/ProjectType' } }
        - name: ownerUserIds
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: '#/components/schemas/Uuid' } }
        - name: actorAccesses
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: '#/components/schemas/ProjectActorAccess' } }
        - { name: updatedSince, in: query, schema: { type: string, format: date-time } }
        - { name: lifecycle, in: query, schema: { $ref: '#/components/schemas/ProjectLifecycleFilter' } }
        - { name: productId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: 按 name、code、id 稳定排序的权限过滤分页。
          content: { application/json: { schema: { $ref: '#/components/schemas/ProjectPage' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      tags: [Projects]
      operationId: createProject
      summary: 原子创建 DRAFT Project
      description: 同一事务创建 Project、ACTIVE owner membership 和模板定义的全部初始 Content。
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreateRequest' }
      responses:
        '201':
          description: Project 已原子创建，或返回同一幂等命令的既有成功结果。
          headers:
            Location: { $ref: '#/components/headers/Location' }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /projects/{projectId}/contents:
    get:
      tags: [Contents]
      operationId: listProjectContents
      summary: 查询 Project Content 目录与固定模板选项
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 当前主体可见的 Content 目录。
          headers: { Cache-Control: { schema: { type: string }, description: 固定为 no-store。 } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProjectContentCatalog' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
    post:
      tags: [Contents]
      operationId: createContent
      summary: 从 Project 固定模板蓝图创建 Content
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ContentCreateRequest' } } }
      responses:
        '201':
          description: Content 已创建。
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location: { $ref: '#/components/headers/Location' }
          content: { application/json: { schema: { $ref: '#/components/schemas/Content' } } }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /projects/{projectId}/contents/{contentId}:
    patch:
      tags: [Contents]
      operationId: updateContent
      summary: 更新工作项类别名称、颜色、顺序或启停状态
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: contentId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ContentUpdateRequest' } } }
      responses:
        '200':
          description: 工作项类别已更新。
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Content' } } }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    delete:
      tags: [Contents]
      operationId: deleteContent
      summary: 软删除从未使用的自定义工作项类别
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: contentId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: 工作项类别已软删除。 }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
  /projects/{projectId}/activity:
    get:
      tags: [Activity]
      operationId: listProjectActivity
      summary: 查询项目动态
      description: 仅返回 V44 切点后的白名单事实；每次请求按当前 Project 可见性重新鉴权。
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: projectId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - $ref: '#/components/parameters/ActivityCursor'
        - $ref: '#/components/parameters/ActivitySize'
        - $ref: '#/components/parameters/ActivityEventType'
        - $ref: '#/components/parameters/ActivityEntityType'
        - $ref: '#/components/parameters/ActivityOccurredFrom'
        - $ref: '#/components/parameters/ActivityOccurredTo'
      responses:
        '200':
          description: 按 occurredAt、id 倒序的项目动态。
          headers: { Cache-Control: { schema: { type: string }, description: 固定为 no-store。 } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ActivityPage' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /work-items/{workItemId}/activity:
    get:
      tags: [Activity]
      operationId: listWorkItemActivity
      summary: 查询事项动态
      description: 软删除事项仍可在当前 Project 可见时读取历史；游标不可跨事项或筛选复用。
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: workItemId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - $ref: '#/components/parameters/ActivityCursor'
        - $ref: '#/components/parameters/ActivitySize'
        - $ref: '#/components/parameters/ActivityEventType'
        - $ref: '#/components/parameters/ActivityEntityType'
        - $ref: '#/components/parameters/ActivityOccurredFrom'
        - $ref: '#/components/parameters/ActivityOccurredTo'
      responses:
        '200':
          description: 按 occurredAt、id 倒序的事项动态。
          headers: { Cache-Control: { schema: { type: string }, description: 固定为 no-store。 } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ActivityPage' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /work-items/{workItemId}/cell-activity:
    get:
      tags: [Activity]
      operationId: listWorkItemCellActivity
      summary: 查询工作项单元格动态
      description: 仅返回专用投影切点后的字段白名单动态；游标绑定事项、全部筛选和首次请求时间锚点。
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: workItemId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - $ref: '#/components/parameters/ActivityCursor'
        - $ref: '#/components/parameters/ActivitySize'
        - name: timeRange
          in: query
          required: false
          schema: { $ref: '#/components/schemas/WorkItemCellActivityTimeRange' }
        - name: actorUserId
          in: query
          required: false
          style: form
          explode: true
          schema: { type: array, uniqueItems: true, items: { $ref: '#/components/schemas/Uuid' } }
        - name: column
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            uniqueItems: true
            items: { $ref: '#/components/schemas/WorkItemCellActivityColumn' }
      responses:
        '200':
          description: 按 occurredAt、id 倒序的工作项单元格动态及完整投影统计。
          headers: { Cache-Control: { schema: { type: string }, description: 固定为 no-store。 } }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemCellActivityPage' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-yumpoo-session
      description: 服务端签发的 HttpOnly、Secure、SameSite=Lax Web Session Cookie。
  schemas:
...
    ActivityPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, historyStartedAt]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ActivityItem' } }
        nextCursor: { type: string, minLength: 1, maxLength: 2048, nullable: true }
        historyStartedAt:
          type: string
          format: date-time
          description: Activity V1 投影开始接收事件的明确切点，更早历史未回填。
    WorkItemCellActivityColumn:
      type: string
      enum: [WORK_ITEM_NAME, ASSIGNEE, STATUS, PRIORITY, DUE_DATE, CONTENT]
...
    WorkItemCellActivityTimeRange:
      type: string
      enum: [TODAY, YESTERDAY, THIS_WEEK, THIS_MONTH, THIS_YEAR]
...
    Uuid:
      type: string
      format: uuid
      description: 不透明 UUID；客户端不得从值中推导业务语义。
...
    ProjectType:
      type: string
      enum: [PRODUCT_DEVELOPMENT, PRE_SALES, IMPLEMENTATION, HYPERCARE]
...
    ProjectLifecycleFilter:
      type: string
      enum: [DRAFT, ACTIVE, ARCHIVED, ALL]
    ProjectActorAccess:
      type: string
      enum: [OWNER, MEMBER, COMPANY_ADMIN]
...
    Project:
      type: object
      additionalProperties: false
      required:
        - id
        - workspaceId
        - code
        - name
        - description
        - projectType
        - lifecycle
        - ownerUserId
        - templateKey
        - templateVersion
        - customerName
        - customerReference
        - deliverySite
        - contactNote
        - rowVersion
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        workspaceId: { $ref: '#/components/schemas/Uuid' }
        code:
          type: string
          minLength: 2
          maxLength: 32
          pattern: '^[A-Z][A-Z0-9_]{1,31}$'
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500, nullable: true }
        projectType: { $ref: '#/components/schemas/ProjectType' }
        lifecycle: { $ref: '#/components/schemas/ProjectLifecycle' }
        ownerUserId: { $ref: '#/components/schemas/Uuid' }
        templateKey: { $ref: '#/components/schemas/ProjectTemplateKey' }
        templateVersion: { type: integer, format: int32, minimum: 1 }
        customerName: { type: string, maxLength: 160, nullable: true }
        customerReference: { type: string, maxLength: 80, nullable: true }
        deliverySite: { type: string, maxLength: 160, nullable: true }
        contactNote: { type: string, maxLength: 500, nullable: true }
        rowVersion: { type: integer, format: int64, minimum: 0, readOnly: true }
    ProjectCreateRequest:
      type: object
      additionalProperties: false
      required:
        - code
        - name
        - projectType
        - ownerUserId
        - templateKey
        - templateVersion
      properties:
        workspaceId: { $ref: '#/components/schemas/Uuid' }
        code:
          type: string
          minLength: 2
          maxLength: 32
          pattern: '^[A-Z][A-Z0-9_]{1,31}$'
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500, nullable: true }
        projectType: { $ref: '#/components/schemas/ProjectType' }
        ownerUserId: { $ref: '#/components/schemas/Uuid' }
        templateKey: { $ref: '#/components/schemas/ProjectTemplateKey' }
        templateVersion: { type: integer, format: int32, minimum: 1 }
        customerName: { type: string, maxLength: 160, nullable: true }
        customerReference: { type: string, maxLength: 80, nullable: true }
        deliverySite: { type: string, maxLength: 160, nullable: true }
        contactNote: { type: string, maxLength: 500, nullable: true }
...
    ProjectPage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProjectSummary' } }
        page: { type: integer, format: int32, minimum: 0 }
        size: { type: integer, format: int32, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, format: int32, minimum: 0 }
...
    Content:
      type: object
      additionalProperties: false
      required: [id, projectId, code, name, colorToken, sortOrder, active,
        protectedContent, inUse, rowVersion, createdAt, createdByUserId,
        updatedAt, updatedByUserId]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        projectId: { $ref: '#/components/schemas/Uuid' }
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]{1,31}$' }
        name: { type: string, minLength: 1, maxLength: 80 }
        colorToken: { $ref: '#/components/schemas/WorkItemLabelColorToken' }
        sortOrder: { type: integer, format: int32, minimum: 1 }
        active: { type: boolean }
        protectedContent: { type: boolean, readOnly: true }
        inUse: { type: boolean, readOnly: true }
        rowVersion: { type: integer, format: int64, minimum: 0, readOnly: true }
        createdAt: { type: string, format: date-time, readOnly: true }
        createdByUserId: { $ref: '#/components/schemas/Uuid' }
        updatedAt: { type: string, format: date-time, readOnly: true }
        updatedByUserId: { $ref: '#/components/schemas/Uuid' }
    ProjectContentCatalog:
      type: object
      additionalProperties: false
      required: [items, rowVersion, etag, canManage]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Content' } }
        rowVersion: { type: integer, format: int64, minimum: 0, readOnly: true }
        etag: { type: string, pattern: '^"[0-9]+"$', readOnly: true }
        canManage: { type: boolean, readOnly: true }
    ContentCreateRequest:
      type: object
      additionalProperties: false
      required: [name, colorToken]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        colorToken: { $ref: '#/components/schemas/WorkItemLabelColorToken' }
    ContentUpdateRequest:
      type: object
      additionalProperties: false
      required: [name, colorToken, active, sortOrder]
      properties:
        name: { type: string, minLength: 1, maxLength: 80, nullable: true }
        colorToken:
          type: string
          allOf: [{ $ref: '#/components/schemas/WorkItemLabelColorToken' }]
          nullable: true
        active: { type: boolean, nullable: true }
        sortOrder: { type: integer, format: int32, minimum: 1, nullable: true }
...
  parameters:
    ActivityCursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 2048 }
    ActivitySize:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    ActivityEventType:
      name: eventType
      in: query
      required: false
      style: form
      explode: true
      schema:
        type: array
        uniqueItems: true
        items: { type: string, minLength: 3, maxLength: 128 }
    ActivityEntityType:
      name: entityType
      in: query
      required: false
      style: form
      explode: true
      schema:
        type: array
        uniqueItems: true
        items: { type: string, minLength: 2, maxLength: 32 }
    ActivityOccurredFrom:
      name: occurredFrom
      in: query
      required: false
      schema: { type: string, format: date-time }
    ActivityOccurredTo:
      name: occurredTo
      in: query
      required: false
      schema: { type: string, format: date-time }
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: 与当前 Session 绑定并由非 HttpOnly Cookie 回传的 CSRF 凭据。
      schema:
        type: string
        minLength: 1
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: |
        单个强 ETag，格式为双引号包围的十进制 rowVersion。弱验证器、`*` 和多个
        ETag 均不属于一期契约。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
      example: '"5"'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端为单一业务意图生成并在重试时复用的 UUID。
      schema:
        type: string
        format: uuid
        maxLength: 64
...
  headers:
    Location:
      description: 跳转目标 URI。
      schema:
        type: string
        format: uri-reference
...
    XRequestId:
      description: 请求关联标识；错误响应中必须与 body.requestId 完全一致。
      schema:
        $ref: '#/components/schemas/RequestId'
    ETag:
      description: 当前资源 rowVersion 对应的单个强 ETag。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
      example: '"5"'
...
  examples:
...
    AuthenticationRequired:
      summary: 会话缺失、未知、过期、主动退出或登录回调统一失败
      externalValue: ../examples/errors/401-authentication-required.json
...
  responses:
...
    InternalError:
      description: 系统错误或高风险操作因 Security Audit 不可写而失败关闭。
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            internalError: { $ref: '#/components/examples/InternalError' }
    MalformedRequest:
      description: 请求 JSON、类型、参数格式或条件头无法解析。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            malformedRequest:
              $ref: '#/components/examples/MalformedRequest'
    AuthenticationRequired:
      description: 未建立有效 CurrentActor。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            authenticationRequired:
              $ref: '#/components/examples/AuthenticationRequired'
    Forbidden:
      description: 可识别账号已禁用，或有效主体缺少可见资源上的动作权限。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            accountDisabled:
              $ref: '#/components/examples/AccountDisabled'
            accessDenied:
              $ref: '#/components/examples/AccessDenied'
    ResourceNotFound:
      description: 资源不存在，或资源存在性对当前主体不可见。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            resourceNotFound:
              $ref: '#/components/examples/ResourceNotFound'
    Conflict:
      description: 幂等、处理中、状态迁移或工时锁定冲突。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            idempotencyKeyReused:
              $ref: '#/components/examples/IdempotencyKeyReused'
            requestInProgress:
              $ref: '#/components/examples/RequestInProgress'
            invalidStateTransition:
              $ref: '#/components/examples/InvalidStateTransition'
            worklogLocked:
              $ref: '#/components/examples/WorklogLocked'
    VersionConflict:
      description: If-Match 与当前可见资源版本不一致。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            versionConflict:
              $ref: '#/components/examples/VersionConflict'
...
    ValidationFailed:
      description: 请求格式正确，但字段、必填项或业务规则不满足。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationFailed:
              $ref: '#/components/examples/ValidationFailed'
...
    PreconditionRequired:
      description: 条件写入缺少必需的 If-Match。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            preconditionRequired:
              $ref: '#/components/examples/PreconditionRequired'
...